Publish the extended fork as a single snapshot
Everything this fork adds over the upstream base in one commit: rendering
parity work (cameras, particles, lighting, models, scripting, text), VRAM
and performance work (BC7 ingest compression, FBO pooling, mip residency),
the two-service architecture with CEF isolation, the daemon command API
with a self-restoring engine, the lwe-ui control panel, and parser and
socket hardening. See README for the full picture.

Development history stays in a private repository; future work lands here
as normal commits. Vendored trees under src/External replace the upstream
submodules.

+++ src/WallpaperEngine/Audio/Drivers/SDLAudioDriver.cpp
@@ -57,10 +57,12 @@ void audio_callback (void* userdata, uint8_t* streamData, int length) {
 		len1 = streamLength;
 	    }
 
-	    // mix the audio
 	    SDL_MixAudioFormat (
 		streamDataPointer, &buffer->audio_buf[buffer->audio_buf_index], driver->getSpec ().format, len1,
-		driver->getApplicationContext ().state.audio.volume
+		static_cast<int> (
+		    static_cast<float> (driver->getApplicationContext ().state.audio.volume)
+		    * buffer->stream->getVolume ()
+		)
 	    );
 
 	    streamLength -= len1;
